feat(cgb): Phase 1 — PPU 出力を RGB555 に統一
- PPU の pixel buffer を [u8]（パレットインデックス）→ [u16]（RGB555）に変更
  - DMG 4 色パレットを DMG_PALETTE として PPU 内に定義し描画時に適用
  - pixel_buffer() の戻り値を &[u16] に変更
- Display トレイトを draw(&[u16]) に変更（platform.rs / NullDisplay）
- host/src/lcd.rs: 4 色マッチを廃止し RGB555→RGB24 変換に置換
- teensy/src/display/mod.rs: PALETTE 参照を廃止し rgb555_to_rgb565() 変換関数を追加
  fill_buffer / draw を &[u16] 対応に変更
- DMG ROM の見た目は変わらず 21 tests all pass

+++ core/src/platform.rs
@@ -1,9 +1,10 @@
 //! プラットフォーム抽象トレイト群（表示・音声・カートリッジバス）。
 //! 入力は [`crate::input::InputSource`] を使う。
 
-/// 160x144 のパレットインデックス(0–3)バッファを表示する。
+/// 160x144 の RGB555 ピクセルバッファを表示する。
+/// bits 0-4=R, bits 5-9=G, bits 10-14=B（GBC ネイティブ形式）
 pub trait Display {
-    fn draw(&mut self, buffer: &[u8]);
+    fn draw(&mut self, buffer: &[u16]);
 }
 
 /// ステレオ f32 サンプルを出力先へ渡す。
@@ -25,7 +26,7 @@ pub trait CartridgeBus {
 pub struct NullDisplay;
 
 impl Display for NullDisplay {
-    fn draw(&mut self, _: &[u8]) {}
+    fn draw(&mut self, _: &[u16]) {}
 }
 
 /// 音声を破棄する no-op 実装。
